+++ app/src/main/java/com/daniyalh/WeatherWiseApp/logic/IWeatherJsonAdapter.java
@@ -1,0 +1,5 @@
+package com.daniyalh.WeatherWiseApp.logic;
+
+public interface IWeatherJsonAdapter {
+    String[] parseWeather(String weatherJSON);
+}
