+++ app/src/main/java/com/daniyalh/WeatherWiseApp/logic/WeatherJsonAdapter.java
@@ -1,6 +1,6 @@
-package com.example.WeatherWiseApp.logic;
+package com.daniyalh.WeatherWiseApp.logic;
 
-import com.example.WeatherWiseApp.logic.exceptions.InvalidJsonParsingException;
+import com.daniyalh.WeatherWiseApp.logic.exceptions.InvalidJsonParsingException;
 import org.json.JSONObject;
 
 public class WeatherJsonAdapter implements IWeatherJsonAdapter {
