+++ app/src/main/java/com/daniyalh/WeatherWiseApp/logic/WeatherManager.java
@@ -1,12 +1,12 @@
-package com.example.WeatherWiseApp.logic;
+package com.daniyalh.WeatherWiseApp.logic;
 
 import android.content.Context;
 
 import com.android.volley.RequestQueue;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
-import com.example.WeatherWiseApp.logic.exceptions.InvalidJsonParsingException;
-import com.example.WeatherWiseApp.objects.City;
+import com.daniyalh.WeatherWiseApp.logic.exceptions.InvalidJsonParsingException;
+import com.daniyalh.WeatherWiseApp.objects.City;
 
 public class WeatherManager implements IWeatherManager {
