+++ app/src/main/java/com/daniyalh/WeatherWiseApp/logic/exceptions/ParsingException.java
@@ -1,4 +1,4 @@
-package com.example.WeatherWiseApp.logic.exceptions;
+package com.daniyalh.WeatherWiseApp.logic.exceptions;
 
 public class ParsingException extends RuntimeException {
     public ParsingException(String errorMessage, Throwable e) {
