+++ app/src/main/java/com/daniyalh/WeatherWiseApp/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.WeatherWiseApp.ui.theme
+package com.daniyalh.WeatherWiseApp.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
