--- app/src/main/java/com/example/WeatherWiseApp/logic/IWeatherJsonAdapter.java
@@ -1,7 +1,0 @@
-package com.example.WeatherWiseApp.logic;
-
-import com.example.WeatherWiseApp.logic.exceptions.InvalidJsonParsingException;
-
-public interface IWeatherJsonAdapter {
-    String[] parseWeather(String weatherJSON);
-}
