--- app/src/main/java/com/example/WeatherWiseApp/logic/IWeatherManager.java
@@ -1,11 +1,0 @@
-package com.example.WeatherWiseApp.logic;
-
-import android.content.Context;
-
-import com.example.WeatherWiseApp.logic.exceptions.InvalidJsonParsingException;
-import com.example.WeatherWiseApp.objects.City;
-
-public interface IWeatherManager {
-    void getWeatherJSON(City city, final IWeatherCallback callback);
-    void setWeather(City city, String weatherJSON) throws InvalidJsonParsingException;
-}
